Fix deprecation warning for implicit int conversion

+++ src/lib/MVC/Symfony/Security/Authentication/RepositoryAuthenticationProvider.php
@@ -114,7 +114,7 @@ class RepositoryAuthenticationProvider extends DaoAuthenticationProvider impleme
 
         $remainingTime = $this->constantAuthTime - (microtime(true) - $startTime);
         if ($remainingTime > 0) {
-            usleep($remainingTime * 1000000);
+            usleep((int)($remainingTime * 1000000));
         } elseif ($this->logger) {
             $this->logger->warning(
                 sprintf(
